fix: address PRRT_kwDOSJAM6s6Zkph6 - split oversized tests
