+++ tests/unit/control/test_worker_coverage_edges_parts/test_worker_coverage_edges_part_002.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import contextlib
 from collections.abc import AsyncIterator
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
@@ -11,7 +10,6 @@ from types import SimpleNamespace
 from unittest.mock import AsyncMock
 
 import pytest
-import structlog
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.common.commands import FakeCommandRunner
@@ -20,7 +18,6 @@ from awf.control.worker import (
     ControlWorker,
     WorkerConfig,
 )
-from awf.control.worker import dispatch_methods as worker_dispatch_methods
 from awf.control.worker import helpers as worker_helpers
 from awf.control.worker import recovery_cooldown as worker_recovery_cooldown
 from awf.control.worker import recovery_stale as worker_recovery_stale
@@ -1484,63 +1481,3 @@ async def test_salvage_monitor_cooldown_active_evicts_expired_entries(
 
     assert not worker._active_salvage_monitor_resume_cooldown_active("ws_expired")  # noqa: SLF001
     assert "ws_expired" not in worker._active_salvage_monitor_resume_cooldowns  # noqa: SLF001
-
-
-@pytest.mark.unit
-async def test_forget_execution_task_ignores_replaced_task(
-    factory: async_sessionmaker[AsyncSession],
-) -> None:
-    """Regression coverage for forget execution task ignores replaced task."""
-    worker = _worker(factory)
-
-    async def _pending() -> None:
-        await asyncio.Event().wait()
-
-    tracked = asyncio.create_task(_pending())
-    stale = asyncio.create_task(_pending())
-    worker._execution_tasks["ws_replaced"] = tracked  # noqa: SLF001
-    worker._execution_task_kinds["ws_replaced"] = (  # noqa: SLF001
-        worker_dispatch_methods._ExecutionTaskKind.READY
-    )
-    try:
-        # A late done-callback from a previously cancelled task must not evict the
-        # task that currently owns the slot (identity guard).
-        worker._forget_execution_task("ws_replaced", stale)  # noqa: SLF001
-        assert worker._execution_tasks["ws_replaced"] is tracked  # noqa: SLF001
-        assert (
-            worker._execution_task_kinds["ws_replaced"]  # noqa: SLF001
-            is worker_dispatch_methods._ExecutionTaskKind.READY
-        )
-
-        # The matching task clears both slot-accounting maps together.
-        worker._forget_execution_task("ws_replaced", tracked)  # noqa: SLF001
-        assert "ws_replaced" not in worker._execution_tasks  # noqa: SLF001
-        assert "ws_replaced" not in worker._execution_task_kinds  # noqa: SLF001
-    finally:
-        for task in (tracked, stale):
-            task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await task
-
-
-@pytest.mark.unit
-async def test_reconcile_drops_monitor_kind_when_task_already_gone(
-    factory: async_sessionmaker[AsyncSession],
-) -> None:
-    """Drop stale monitor-resume kind entries when the workspace row is already gone."""
-    worker = _worker(factory)
-    # Simulate a monitor resume that finished concurrently during the status load:
-    # the kind entry lingers while its task is already gone and the workspace row
-    # is absent (so it is no longer monitoring_pr). Reconcile should drop the stale
-    # kind entry without logging a cancellation or raising.
-    worker._execution_task_kinds["ws_vanished"] = (  # noqa: SLF001
-        worker_dispatch_methods._ExecutionTaskKind.MONITOR_RESUME
-    )
-
-    with structlog.testing.capture_logs() as captured:
-        await worker._reconcile_stale_monitor_execution_tasks()  # noqa: SLF001
-
-    assert "ws_vanished" not in worker._execution_task_kinds  # noqa: SLF001
-    assert not any(
-        event.get("event") == "worker.stale_monitor_execution_task_cancelled" for event in captured
-    )
