+++ tests/unit/db/test_workspace_repository_parts/test_workspace_repository_part_002.py
@@ -1465,45 +1465,6 @@ class TestOwnedPathOverlapLookup:
         ]
 
     @pytest.mark.unit
-    async def test_non_postgres_get_for_update_omits_row_lock(self) -> None:
-        """Verify non-Postgres get-for-update remains a plain select."""
-        session = _RecordingSchedulerSession("sqlite", values=["ws_unlocked"])
-        repo = WorkspaceRepository(session, dialect_name="sqlite")  # type: ignore[arg-type]
-
-        locked = await repo.get_for_update("ws_unlocked", skip_locked=True)
-
-        assert locked == "ws_unlocked"
-        assert len(session.executed) == 1
-        sql = str(
-            session.executed[0].compile(  # type: ignore[attr-defined]
-                dialect=sqlite.dialect(),
-                compile_kwargs={"literal_binds": True},
-            )
-        )
-        assert "FOR UPDATE" not in sql
-        assert "SKIP LOCKED" not in sql
-        assert "workspaces.id = 'ws_unlocked'" in sql
-
-    @pytest.mark.unit
-    async def test_postgres_get_for_update_locks_workspace_row(self) -> None:
-        """Verify Postgres get-for-update locks the workspace row."""
-        session = _RecordingSchedulerSession("postgresql", values=["ws_locked"])
-        repo = WorkspaceRepository(session, dialect_name="postgresql")  # type: ignore[arg-type]
-
-        locked = await repo.get_for_update("ws_locked")
-
-        assert locked == "ws_locked"
-        assert len(session.executed) == 1
-        sql = str(
-            session.executed[0].compile(  # type: ignore[attr-defined]
-                dialect=postgresql.dialect(),
-                compile_kwargs={"literal_binds": True},
-            )
-        )
-        assert "FOR UPDATE" in sql
-        assert "workspaces.id = 'ws_locked'" in sql
-
-    @pytest.mark.unit
     async def test_session_info_dialect_drives_scheduler_locking(self) -> None:
         """Verify session dialect metadata enables scheduler locking."""
         session = _RecordingSchedulerSession(
